+++ src/js/ui/DOM/Library/basic/index.js
@@ -1,9 +1,11 @@
 import HTML from "./html"
 import SVG from "./svg"
 import VideoBlock from "./video"
+import Link from "./link"
 
 export {
     HTML,
     SVG,
     VideoBlock,
+    Link,
 }