+++ changelog.md
@@ -1,3 +1,4 @@
 # 0.4.0 11.03.2019
 ## General
 + App changelog generation
++ Last version changelog loading for update popup